Add docs and move conversion to its own module
